pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aperion-event-bus"
version = "0.3.0"
description = "The Nervous System - A standalone event bus with correlation ID tracking, audit logging, and PII safety"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Aperion Team" }
]
keywords = ["event-bus", "pub-sub", "observability", "telemetry", "correlation-id", "opentelemetry"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

dependencies = []

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.4",
    "mypy>=1.9",
    "fastapi>=0.100.0",
    "httpx>=0.25.0",
    "starlette>=0.27.0",
]
# OpenTelemetry integration (optional)
otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]
# Prometheus metrics export (optional)
prometheus = [
    "prometheus-client>=0.19.0",
]
# All observability features
observability = [
    "aperion-event-bus[otel,prometheus]",
]

[project.urls]
Homepage = "https://github.com/aperion/aperion-event-bus"
Documentation = "https://github.com/aperion/aperion-event-bus#readme"
Repository = "https://github.com/aperion/aperion-event-bus"

[tool.hatch.build.targets.wheel]
packages = ["src/aperion_event_bus"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "anyio: mark test to run with anyio",
    "slow: Long-running stability tests (60+ seconds)",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
